requirements.txt
python-telegram-bot==22.2
requests==2.32.4
pandas==2.3.1
numpy==2.3.1
python-dotenv==1.1.1